feat(payroll): implement safe cancellation path for pending payroll runs (#198)
- Add finalize_payroll_run to complete the two-step prepare-to-finalize flow
- Modify cancel_payroll_run with mandatory reason parameter for audit trails
- Remove pause check from cancel_payroll_run (high-priority escape hatch)
- Add guard against canceling already-finalized payroll runs
- Consume draft commitment during finalization for one-time-use integrity
- Remove broken libmain.rs duplicate file
- Add 9 comprehensive tests covering finalize, pause+escape, state guards
