+++ contracts/payroll/src/lib.rs
@@ -699,15 +699,25 @@ impl Payroll {
         e.storage().persistent().get(&DataKey::PendingRun(run_id))
     }
 
-    /// Cancel a pending payroll run without executing any payments.
+    /// Finalize a pending payroll run, executing payments and creating a
+    /// permanent `PayrollRun` record (issue #198).
     ///
-    /// Only the admin may cancel. The cancellation frees the nonce for reuse
-    /// (actually, the nonce is marked as consumed, so it cannot be reused).
-    /// No funds are transferred; this is a pure cleanup operation.
+    /// Only the admin may finalize. The caller must supply the same proofs,
+    /// amounts, and employees that were validated during `prepare_payroll_run`.
+    /// The pending run must exist and its metadata (total_amount, employee_count)
+    /// must match the supplied batch.
     ///
-    /// Cancellation emits an event for audit trails. Finalized runs cannot be
-    /// cancelled retroactively.
-    pub fn cancel_payroll_run(e: Env, admin: Address, run_id: u64) {
+    /// On success the pending run is removed and a completed `PayrollRun` is
+    /// stored. If proofs fail verification the pending run is preserved so the
+    /// admin can still cancel or retry.
+    pub fn finalize_payroll_run(
+        e: Env,
+        admin: Address,
+        run_id: u64,
+        proofs: Vec<BytesN<256>>,
+        amounts: Vec<i128>,
+        employees: Vec<Address>,
+    ) {
         Self::require_not_paused(&e);
         let addrs: ContractAddresses = e
             .storage()
@@ -726,13 +736,167 @@ impl Payroll {
             .get(&pending_key)
             .expect("Pending run not found");
 
+        let count = proofs.len();
+        if amounts.len() != count || employees.len() != count {
+            panic!("Array length mismatch");
+        }
+        assert!(count <= MAX_BATCH, "Batch too large");
+
+        // Verify employee count matches the prepared run.
+        if count as u32 != pending_run.employee_count {
+            panic!(
+                "Employee count mismatch: prepared {} but batch has {}",
+                pending_run.employee_count, count
+            );
+        }
+
+        // Verify total spend matches the prepared run.
+        let mut total: i128 = 0;
+        for i in 0..count {
+            let amt = amounts.get(i).unwrap();
+            if amt <= 0 {
+                panic!("Amount must be positive");
+            }
+            total += amt;
+        }
+        if total != pending_run.total_amount {
+            panic!(
+                "Total spend mismatch: prepared {} but batch totals {}",
+                pending_run.total_amount, total
+            );
+        }
+
+        // Consume the draft commitment if one was bound to this run (#102).
+        if pending_run.draft_hash != BytesN::from_array(&e, &[0u8; 32]) {
+            let commit_key = DataKey::DraftCommitment(pending_run.draft_hash.clone());
+            if e.storage().persistent().has(&commit_key) {
+                e.storage().persistent().remove(&commit_key);
+            }
+        }
+
+        // Verify proofs and execute payments.
+        let verifier = ProofVerifierClient::new(&e, &addrs.verifier);
+        let commitment_client =
+            SalaryCommitmentContractClient::new(&e, &addrs.commitment);
+        let token_client = soroban_token::Client::new(&e, &addrs.token);
+
+        for i in 0..count {
+            let proof = proofs.get(i).unwrap();
+            let amount = amounts.get(i).unwrap();
+            let employee = employees.get(i).unwrap();
+
+            let commitment_struct = commitment_client.get_commitment(&employee);
+            let commitment = commitment_struct.commitment;
+
+            let mut nullifier_arr = [0u8; 32];
+            nullifier_arr[0] = (i % 256) as u8;
+            nullifier_arr[1] = (i / 256) as u8;
+            let nullifier = BytesN::from_array(&e, &nullifier_arr);
+            let recipient_hash = BytesN::from_array(&e, &[0u8; 32]);
+
+            let mut public_inputs = Vec::new(&e);
+            public_inputs.push_back(commitment.clone());
+            public_inputs.push_back(nullifier.clone());
+            public_inputs.push_back(recipient_hash.clone());
+
+            let ok = verifier.verify_payment_proof(&proof, &public_inputs);
+            if !ok {
+                panic!("Invalid payment proof for employee {}", i);
+            }
+
+            commitment_client.record_nullifier(&nullifier);
+
+            token_client.transfer(&addrs.treasury, &employee, &amount);
+
+            commitment_client.lock_commitment_updates(&employee);
+
+            e.events().publish(
+                (
+                    symbol_short!("payroll"),
+                    Symbol::new(&e, "payment_executed"),
+                ),
+                (employee.clone(), amount),
+            );
+        }
+
+        // Remove the pending run and create the permanent record.
+        e.storage().persistent().remove(&pending_key);
+
+        let run = PayrollRun {
+            run_id,
+            executed_at: e.ledger().timestamp(),
+            admin: addrs.admin.clone(),
+            total_amount: pending_run.total_amount,
+            employee_count: pending_run.employee_count,
+            draft_hash: pending_run.draft_hash.clone(),
+            nonce: pending_run.nonce.clone(),
+            reconciliation_status: ReconciliationStatus::Unreconciled,
+            metadata_hash: BytesN::from_array(&e, &[0u8; 32]),
+        };
+        e.storage()
+            .persistent()
+            .set(&DataKey::PayrollRun(run_id), &run);
+
+        e.events().publish(
+            (
+                symbol_short!("payroll"),
+                Symbol::new(&e, "run_finalized"),
+            ),
+            (run_id, pending_run.total_amount),
+        );
+    }
+
+    /// Cancel a pending payroll run without executing any payments (issue #198).
+    ///
+    /// Only the admin may cancel. The `reason` is recorded in the event for
+    /// audit trails. No funds are transferred; this is a pure cleanup operation.
+    ///
+    /// Finalized runs cannot be cancelled retroactively. The run nonce remains
+    /// permanently spent after cancellation (one-time-use for audit integrity).
+    ///
+    /// This function serves as a high-priority escape hatch: it deliberately
+    /// does NOT require the system to be unpaused. An admin who can pause the
+    /// system can also cancel a pending run while paused, enabling rapid
+    /// intervention when a run is discovered to be unsafe.
+    pub fn cancel_payroll_run(e: Env, admin: Address, run_id: u64, reason: Symbol) {
+        let addrs: ContractAddresses = e
+            .storage()
+            .persistent()
+            .get(&DataKey::Addresses)
+            .expect("Not initialized");
+        if admin != addrs.admin {
+            panic!("Unauthorized");
+        }
+        admin.require_auth();
+
+        let pending_key = DataKey::PendingRun(run_id);
+
+        // Guard: reject cancellation if a finalized PayrollRun already exists
+        // for this run_id (completed via finalize_payroll_run or
+        // batch_process_payroll).
+        if e.storage()
+            .persistent()
+            .has(&DataKey::PayrollRun(run_id))
+        {
+            panic!("Cannot cancel a finalized payroll run");
+        }
+
+        let pending_run: PendingPayrollRun = e
+            .storage()
+            .persistent()
+            .get(&pending_key)
+            .expect("Pending run not found");
+
         // Remove the pending run from storage
         e.storage().persistent().remove(&pending_key);
 
-        // Emit cancellation event
+        // Emit cancellation event with reason for audit trail
         e.events().publish(
-            (symbol_short!("payroll"), Symbol::new(&e, "run_cancelled")),
-            (run_id, pending_run.total_amount),
+            (
+                symbol_short!("payroll"),
+                Symbol::new(&e, "run_cancelled"),
+            ),
+            (run_id, pending_run.total_amount, reason),
         );
     }
 
@@ -2248,7 +2412,8 @@ mod tests {
         );
 
         // Cancel the pending run
-        payroll_client.cancel_payroll_run(&admin, &run_id);
+        let reason = Symbol::new(&env, "test_cancel");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
 
         // Verify it's no longer pending
         assert!(payroll_client.get_pending_run(&run_id).is_none());
@@ -2272,7 +2437,8 @@ mod tests {
         );
 
         let non_admin = Address::generate(&env);
-        payroll_client.cancel_payroll_run(&non_admin, &run_id);
+        let reason = Symbol::new(&env, "attack");
+        payroll_client.cancel_payroll_run(&non_admin, &run_id, &reason);
     }
 
     #[test]
@@ -2282,7 +2448,8 @@ mod tests {
         let (payroll_client, admin, _treasury, _treasury_owner, _employee) =
             setup_simple_payroll(&env);
 
-        payroll_client.cancel_payroll_run(&admin, &999u64);
+        let reason = Symbol::new(&env, "no_such_run");
+        payroll_client.cancel_payroll_run(&admin, &999u64, &reason);
     }
 
     // ── Issue #177: payroll run metadata hash checks ──────────────────────────
@@ -2396,8 +2563,9 @@ mod tests {
             &None,
         );
 
-        payroll_client.cancel_payroll_run(&admin, &run_id);
-        payroll_client.cancel_payroll_run(&admin, &run_id);
+        let reason = Symbol::new(&env, "double_cancel");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
     }
 
     #[test]
@@ -2418,7 +2586,8 @@ mod tests {
         );
 
         assert!(payroll_client.get_pending_run(&run_id).is_some());
-        payroll_client.cancel_payroll_run(&admin, &run_id);
+        let reason = Symbol::new(&env, "test_cleanup");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
 
         assert!(payroll_client.get_pending_run(&run_id).is_none());
     }
@@ -2663,7 +2832,8 @@ mod tests {
 
         // Failed cancel (wrong caller) should not affect the pending run
         let attacker = Address::generate(&env);
-        let cancel_result = payroll_client.try_cancel_payroll_run(&attacker, &run_id);
+        let reason = Symbol::new(&env, "attack");
+        let cancel_result = payroll_client.try_cancel_payroll_run(&attacker, &run_id, &reason);
         assert!(cancel_result.is_err());
 
         // Pending run should still exist
@@ -2918,9 +3088,10 @@ mod tests {
             &None,
         );
 
-        payroll_client.cancel_payroll_run(&admin, &run_id);
+        let reason = Symbol::new(&env, "settlement_guard");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
 
-        let result = payroll_client.try_cancel_payroll_run(&admin, &run_id);
+        let result = payroll_client.try_cancel_payroll_run(&admin, &run_id, &reason);
         assert!(result.is_err());
     }
 
@@ -2948,6 +3119,284 @@ mod tests {
         assert!(result.is_err());
 
         let result = payroll_client.try_finalize_run_draft(&admin, &draft_id);
+        assert!(result.is_err());
+    }
+
+    // ── Issue #198: safe cancellation path ───────────────────────────────────
+
+    #[test]
+    fn test_finalize_payroll_run_success() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 60),
+            &None,
+        );
+
+        // Finalize should execute payments and create a PayrollRun
+        payroll_client.finalize_payroll_run(&admin, &run_id, &proofs, &amounts, &employees);
+
+        // Pending run should be gone
+        assert!(payroll_client.get_pending_run(&run_id).is_none());
+
+        // Completed run should exist
+        let run = payroll_client.get_payroll_run(&run_id);
+        assert_eq!(run.run_id, run_id);
+        assert_eq!(run.total_amount, 1000);
+        assert_eq!(run.employee_count, 1);
+    }
+
+    #[test]
+    #[should_panic(expected = "Employee count mismatch")]
+    fn test_finalize_payroll_run_rejects_employee_count_mismatch() {
+        let env = Env::default();
+        env.mock_all_auths();
+
+        let verifier_id = env.register_contract(None, ProofVerifier);
+        let verifier_client = ProofVerifierClient::new(&env, &verifier_id);
+        verifier_client.init_verifier_admin(&Address::generate(&env));
+        verifier_client.initialize_verifier(&mock_vk(&env));
+
+        let commitment_id = env.register_contract(None, SalaryCommitmentContract);
+        let commitment_client = SalaryCommitmentContractClient::new(&env, &commitment_id);
+        commitment_client.init_commitment_admin(&Address::generate(&env));
+
+        let token_id = env.register_contract(None, Token);
+        let token_client = TokenClient::new(&env, &token_id);
+
+        let payroll_id = env.register_contract(None, Payroll);
+        let payroll_client = PayrollClient::new(&env, &payroll_id);
+
+        let treasury = Address::generate(&env);
+        let admin = Address::generate(&env);
+        let treasury_owner = Address::generate(&env);
+        token_client.mint(&treasury, &1_000_000i128);
+        payroll_client.initialize(&admin, &token_id, &verifier_id, &commitment_id, &treasury, &treasury_owner);
+        commitment_client.set_payroll_operator(&payroll_id);
+
+        let employee = Address::generate(&env);
+        commitment_client.store_commitment(&employee, &BytesN::from_array(&env, &[0u8; 32]));
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs, &amounts, &employees, &1000, &test_nonce(&env, 61), &None,
+        );
+
+        // Create a batch with 2 employees (prepared run has 1)
+        let employee2 = Address::generate(&env);
+        commitment_client.store_commitment(&employee2, &BytesN::from_array(&env, &[0u8; 32]));
+        let mut proofs2 = Vec::new(&env);
+        proofs2.push_back(mock_proof(&env));
+        proofs2.push_back(mock_proof(&env));
+        let mut amounts2 = Vec::new(&env);
+        amounts2.push_back(500i128);
+        amounts2.push_back(500i128);
+        let mut employees2 = Vec::new(&env);
+        employees2.push_back(employee.clone());
+        employees2.push_back(employee2.clone());
+
+        payroll_client.finalize_payroll_run(&admin, &run_id, &proofs2, &amounts2, &employees2);
+    }
+
+    #[test]
+    #[should_panic(expected = "Total spend mismatch")]
+    fn test_finalize_payroll_run_rejects_total_mismatch() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 62),
+            &None,
+        );
+
+        // Different total
+        let (proofs2, amounts2, employees2) = single_payment_batch(&env, &employee, 2000);
+        payroll_client.finalize_payroll_run(&admin, &run_id, &proofs2, &amounts2, &employees2);
+    }
+
+    #[test]
+    #[should_panic(expected = "Unauthorized")]
+    fn test_finalize_payroll_run_rejects_non_admin() {
+        let env = Env::default();
+        let (payroll_client, _admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 63),
+            &None,
+        );
+
+        let attacker = Address::generate(&env);
+        payroll_client.finalize_payroll_run(&attacker, &run_id, &proofs, &amounts, &employees);
+    }
+
+    #[test]
+    fn test_cancel_payroll_run_while_paused() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let pm_id = env.register_contract(None, PauseManager);
+        let pm_client = PauseManagerClient::new(&env, &pm_id);
+        let operator = Address::generate(&env);
+        pm_client.initialize(&operator);
+        payroll_client.set_pause_manager(&pm_id);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 64),
+            &None,
+        );
+
+        // Pause the system
+        pm_client.pause();
+
+        // Cancellation should still work (escape hatch)
+        let reason = Symbol::new(&env, "unsafe_amounts");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
+        assert!(payroll_client.get_pending_run(&run_id).is_none());
+    }
+
+    #[test]
+    #[should_panic(expected = "Cannot cancel a finalized payroll run")]
+    fn test_cannot_cancel_finalized_run() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 65),
+            &None,
+        );
+
+        // Finalize first
+        payroll_client.finalize_payroll_run(&admin, &run_id, &proofs, &amounts, &employees);
+
+        // Then try to cancel the finalized run
+        let reason = Symbol::new(&env, "too_late");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
+    }
+
+    #[test]
+    fn test_cancel_then_finalize_fails() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1000);
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1000,
+            &test_nonce(&env, 66),
+            &None,
+        );
+
+        // Cancel first
+        let reason = Symbol::new(&env, "config_error");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
+        assert!(payroll_client.get_pending_run(&run_id).is_none());
+
+        // Then try to finalize the cancelled run
+        let result = payroll_client.try_finalize_payroll_run(
+            &admin, &run_id, &proofs, &amounts, &employees,
+        );
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_prepare_finalize_full_cycle() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 2000);
+
+        // Step 1: Prepare
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &2000,
+            &test_nonce(&env, 67),
+            &None,
+        );
+        assert!(payroll_client.get_pending_run(&run_id).is_some());
+
+        // Step 2: Finalize (executes payments)
+        payroll_client.finalize_payroll_run(&admin, &run_id, &proofs, &amounts, &employees);
+
+        // Verify final state
+        assert!(payroll_client.get_pending_run(&run_id).is_none());
+        let run = payroll_client.get_payroll_run(&run_id);
+        assert_eq!(run.total_amount, 2000);
+        assert_eq!(run.employee_count, 1);
+    }
+
+    #[test]
+    fn test_prepare_cancel_full_cycle() {
+        let env = Env::default();
+        let (payroll_client, admin, _treasury, _treasury_owner, employee) =
+            setup_simple_payroll(&env);
+
+        let (proofs, amounts, employees) = single_payment_batch(&env, &employee, 1500);
+
+        // Step 1: Prepare
+        let run_id = payroll_client.prepare_payroll_run(
+            &proofs,
+            &amounts,
+            &employees,
+            &1500,
+            &test_nonce(&env, 68),
+            &None,
+        );
+        assert!(payroll_client.get_pending_run(&run_id).is_some());
+
+        // Step 2: Cancel with audit reason
+        let reason = Symbol::new(&env, "wrong_period");
+        payroll_client.cancel_payroll_run(&admin, &run_id, &reason);
+
+        // Verify clean state
+        assert!(payroll_client.get_pending_run(&run_id).is_none());
+
+        // Nonce is consumed (can't reuse for another run)
+        let (proofs2, amounts2, employees2) = single_payment_batch(&env, &employee, 1500);
+        let result = payroll_client.try_batch_process_payroll(
+            &proofs2,
+            &amounts2,
+            &employees2,
+            &1500,
+            &test_nonce(&env, 68),  // same seed = same nonce
+            &None,
+        );
         assert!(result.is_err());
     }
 }
